Fix dependency between phasar_llvm_controlflow and phasar_llvm_pointer
