+++ include/phasar/PhasarLLVM/ControlFlow/Resolver/Resolver.h
@@ -18,16 +18,15 @@
 #define PHASAR_PHASARLLVM_CONTROLFLOW_RESOLVER_RESOLVER_H_
 
 #include "phasar/PhasarLLVM/Pointer/LLVMAliasInfo.h"
+#include "phasar/PhasarLLVM/Utils/VirtualCallUtils.h"
 #include "phasar/Utils/MaybeUniquePtr.h"
 
 #include "llvm/ADT/DenseSet.h"
 #include "llvm/ADT/SmallVector.h"
-#include "llvm/IR/DerivedTypes.h"
 
 #include <memory>
 #include <optional>
 #include <string>
-#include <tuple>
 
 namespace llvm {
 class Instruction;
@@ -42,26 +41,6 @@ class LLVMVFTableProvider;
 class DIBasedTypeHierarchy;
 enum class CallGraphAnalysisType;
 
-/// Assuming that `CallSite` is a virtual call through a vtable, retrieves the
-/// index in the vtable of the virtual function called.
-[[nodiscard]] std::optional<unsigned>
-getVFTIndex(const llvm::CallBase *CallSite);
-
-/// Similar to getVFTIndex(), but also returns a pointer to the vtable
-[[nodiscard]] std::optional<std::pair<const llvm::Value *, uint64_t>>
-getVFTIndexAndVT(const llvm::CallBase *CallSite);
-
-/// Detects the pattern \c call(load(GEP(base, const_indices...))) with a
-/// typed (>=3-operand) GEP, i.e. an indirect call through a struct function
-/// pointer field. Distinct from the 2-operand raw-pointer C++ vptr case
-/// handled by \c getVFTIndexAndVT.
-///
-/// Returns \c {base_ptr, all_GEP_indices, gep_source_elem_ty} on match,
-/// or \c std::nullopt otherwise.
-[[nodiscard]] std::optional<std::tuple<
-    const llvm::Value *, llvm::SmallVector<uint64_t, 3>, llvm::Type *>>
-getStructVCallInfo(const llvm::CallBase *CallSite);
-
 /// Assuming that `CallSite` is a call to a non-static member function,
 /// retrieves the type of the receiver. Returns nullptr, if the receiver-type
 /// could not be extracted
@@ -75,12 +54,6 @@ getReceiverType(const llvm::CallBase *CallSite);
     const psr::LLVMVFTableProvider &VTP, const llvm::DIType *ReceiverType);
 
 [[nodiscard]] std::string getReceiverTypeName(const llvm::CallBase *CallSite);
-
-/// Checks whether the signature of `DestFun` matches the required withature of
-/// `CallSite`, such that `DestFun` qualifies as callee-candidate, if `CallSite`
-/// is an indirect/virtual call.
-[[nodiscard]] bool isConsistentCall(const llvm::CallBase *CallSite,
-                                    const llvm::Function *DestFun);
 
 [[nodiscard]] bool isVirtualCall(const llvm::Instruction *Inst,
                                  const LLVMVFTableProvider &VTP);
